requirement.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
pydantic==2.6.2
pyrebase4==4.6.2
firebase-admin==7.5.2

